fix: resolve remaining CI compilation errors
- OsuApiClient: header() -> setHeader(), fix InputStream.readBytes()
- AudioSystem: split playHitsound into playColumnHitsound/playJudgementHitsound to resolve overload ambiguity
- GameEngine: add missing constantSpeed field, use renamed audio methods
- ManiaDifficulty: fix Float.pow() to use Double.pow()
- OsuManiaGame: add missing LoadingScreen import
- LoadingScreen: fix lambda variable capture issue
- BeatmapImportScreen: remove unused BeatmapManager/ApiBeatmapSet references

+++ core/src/main/java/com/leaf/osumania/beatmap/ManiaDifficulty.kt
@@ -13,7 +13,7 @@ private const val SECTION_LENGTH = 400f
 private const val RELEASE_THRESHOLD = 30f
 
 private fun applyDecay(value: Float, deltaTime: Float, decayBase: Float): Float {
-    return value * decayBase.pow((deltaTime / 1000f).toDouble()).toFloat()
+    return value * decayBase.toDouble().pow((deltaTime / 1000f).toDouble()).toFloat()
 }
 
 private fun evaluateIndividualDifficulty(holdFactor: Float): Float {
